+++ src/main/java/com/example/its/web/issue/IssueController.java
@@ -1,4 +1,4 @@
-package com.example.its.web;
+package com.example.its.web.issue;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -12,7 +12,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import com.example.its.domain.issue.IssueService;
 
 import lombok.RequiredArgsConstructor;
-import templates.issues.IssueForm;
 
 @Controller
 @RequestMapping("/issues")
